Read render visibility from the scene store in the RHI backend
Phase 3 of the scene visibility store plan (#579): the render thread
stops keeping hand-synced visibility copies and reads one immutable
cwVisibilitySnapshot instead, captured per sync at the barrier.

- cwRhiScene::synchroize captures scene->visibility()->snapshot() into
  cwRhiFrameRenderer; gatherScene's object gate, anyCloudVisible, and
  atlasIncompatibleVisibleObjectIds read it by id. The per-job
  hiddenObjectIds overlay still ANDs on top, now tested directly by id
  (objects carry their own ids, so the id->pointer resolve pass is gone).
- cwRHIObject: retire the m_isVisible/setVisible mirror; add
  renderObjectId(), stamped in registerRenderObject. GatherContext
  carries a snapshot pointer for sub-item readers. Redundant per-object
  isVisible() short-circuits removed from subclass gather()s;
  usesPointCloudPass() reports geometry only.
- cwRhiTexturedItems: gather reads subVisible(renderObjectId(), itemId)
  from the snapshot. Retire PendingCommand::UpdateVisiblity, the
  RHI-side Item::visible, and the whole-object m_visible copy;
  setItemVisible publishes to the store and schedules a sync instead of
  queueing a command. Dead commented render() body deleted.
- cwRHILinePlot: updateVisibilityBuffer uploads from the snapshot mask,
  gated on (entryVersion, vertexCount); a null mask synthesizes the
  all-visible buffer. Retire the synced cwTracked copy. Front-end
  cwRenderLinePlot::m_visibility de-tracked to a plain QVector; kSubId
  promoted to the header. Shader and vertex widening unchanged.
- Add QDebug operator<<(cwRenderObjectId); seven intersecter log casts
  simplified (the section 3 fold-in follow-up).
- Tests: the two whole-object characterization tests re-express their
  assertions against the frame's snapshot via a new
  CwRhiSceneTestAccess::visibilitySnapshot accessor - the re-expression
  the Phase 0 header reserved. Pick-visibility suites untouched.
- Plan: Phase 3 landing note + gate results; stale sections 0/1.3/6 and
  Q4 corrected; Phase 4 re-scoped to a cleanup sweep (no legacy publish
  path survived Phase 3); five Phase 3 review follow-ups recorded in
  section 3.

Reviewed by the five-lens pre-commit review: zero defects.

Built with cmake --build build/Qt_6_11_1_for_macOS-Debug --target
cavewhere-test cavewhere-qml-test CaveWhere. Full C++ suite green
(312,744 assertions in 1,501 cases); full QML suite green (685 passed,
0 failed, 26 skipped; one first-run NoteOutsideScrapWarning failure
confirmed an ordering flake - passed isolated and in a second full
run). Manual keyword-filter smoke in-app passed.

+++ cavewherelib/src/cwGeometryItersecter.cpp
@@ -685,7 +685,7 @@ void cwGeometryItersecter::clear(cwRenderObjectId parentId)
         }
     }
     qCDebug(lcPick).nospace()
-        << "clear(parent=" << static_cast<quint64>(parentId) << ") — erased "
+        << "clear(parent=" << parentId << ") — erased "
         << erased << " Nodes; remaining=" << Nodes.size();
     // Most render objects leaving the scene registered no pick geometry at
     // all — and a rebuild that would find nothing changed still cancels and
@@ -708,7 +708,7 @@ void cwGeometryItersecter::removeObject(const Key &objectKey)
 {
     if (eraseNodeIfPresent(objectKey)) {
         qCDebug(lcPick).nospace()
-            << "removeObject {parent=" << static_cast<quint64>(objectKey.parentId)
+            << "removeObject {parent=" << objectKey.parentId
             << ", id=" << objectKey.id << "}"
             << " Nodes.size after=" << Nodes.size();
         scheduleTopLevelRebuild();
@@ -725,12 +725,12 @@ void cwGeometryItersecter::setModelMatrix(const Key &objectKey, const QMatrix4x4
     auto iter = findNode(objectKey);
     if (iter == Nodes.end()) {
         qCDebug(lcPick).nospace()
-            << "setModelMatrix {parent=" << static_cast<quint64>(objectKey.parentId)
+            << "setModelMatrix {parent=" << objectKey.parentId
             << ", id=" << objectKey.id << "} — Key not in Nodes; no-op";
         return;
     }
     qCDebug(lcPick).nospace()
-        << "setModelMatrix {parent=" << static_cast<quint64>(objectKey.parentId)
+        << "setModelMatrix {parent=" << objectKey.parentId
         << ", id=" << objectKey.id << "} — top-level rebuild only"
         << " (sub-BVH cache preserved)";
     iter->Object.setModelMatrix(modelMatrix);
@@ -1651,7 +1651,7 @@ cwRayHit cwGeometryItersecter::intersectsDetailed(const QRay3D &ray, const cwPic
                 const Key subKey = sb.object.key();
                 qCDebug(lcPick).nospace()
                     << "    [" << i << "] " << QLatin1String(parentClass)
-                    << "#" << static_cast<quint64>(subKey.parentId)
+                    << "#" << subKey.parentId
                     << " type=" << cwGeometry::typeName(sb.object.geometry().type())
                     << " prims=" << sb.primitives.size()
                     << " visible="
@@ -1985,7 +1985,7 @@ void cwGeometryItersecter::scheduleObjectRebuild(const Key& key)
     m_dirtyKeys.insert(key);
     invalidatePublishedSlot(key);
     qCDebug(lcPick).nospace()
-        << "scheduleObjectRebuild {parent=" << static_cast<quint64>(key.parentId)
+        << "scheduleObjectRebuild {parent=" << key.parentId
         << ", id=" << key.id << "} — sub-BVH invalidated; "
         << m_subBvhs.size() << " other sub-BVH(s) still cached, "
         << m_dirtyKeys.size() << " dirty";
@@ -2510,8 +2510,7 @@ QFuture<void> cwGeometryItersecter::launchBuildJob()
                     ? built[i]->primitives.size()
                     : qsizetype(0);
                 qCDebug(lcPick).nospace()
-                    << "  task[" << i << "] {parent="
-                    << static_cast<quint64>(tasks[i].key.parentId)
+                    << "  task[" << i << "] {parent=" << tasks[i].key.parentId
                     << ", id=" << tasks[i].key.id << "}"
                     << " prims=" << prims
                     << " ms=" << taskMs[i];
